Add files via upload
I created this Game as a final project for the course Matlab and its Applications in Engineering that I had in my year abroad in Barcelona at Universitat Politècnica de Catalunya. I uploaded the full archive. Inside there are the code, the demo and the documentation of the project